1. np_signal_extractor_energy_entropy -- bugfix.

+++ dexterous_bioprosthesis_2021_raw_datasets/set_creators/np_signal_extractors/np_signal_extractor.py
@@ -19,7 +19,7 @@ class NPSignalExtractor(abc.ABC):
 
     def _sanitize_output(self, X):
         if self.sanitize_output:
-            return np.nan_to_num(X, nan=0.0)
+            return np.nan_to_num(X, nan=0.0).astype(X.dtype)
 
         return X
